requirements.txt
aggdraw==1.3.19
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
biplist==1.0.3
certifi==2024.8.30
click==8.1.7
coverage==7.6.7
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
imageio==2.35.1
iniconfig==2.0.0
Jinja2==3.1.4
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
networkx==3.3
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psd-tools==1.9.34
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyprocreate==0.1.0
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lzo==1.15
python-multipart==0.0.10
PyYAML==6.0.2
rich==13.8.1
scikit-image==0.24.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.5
tifffile==2024.9.20
tomli==2.1.0
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
